events component script moved to EVENT.api
